requirements.txt
Django>=4.2,<4.3
djangorestframework>=3.14,<3.15
schedule>=1.2.0
